+++ packages/components-native/src/__mocks__/__mocks.ts
@@ -1,4 +1,5 @@
 /* eslint-disable import/no-internal-modules */
+import { MissingTranslationError } from "react-intl";
 import mockRNLocalize from "react-native-localize/mock";
 
 jest.mock("react-native-localize", () => ({
@@ -22,5 +23,22 @@ jest.mock("react-native-reanimated", () => {
     FadeIn: {
       duration: () => undefined,
     },
+export const MOCK_INTL_CONFIG = {
+  locale: "en-US",
+  onError: (error: MissingTranslationError): void => {
+    if (error.code === "MISSING_TRANSLATION") {
+      return;
+    }
+    throw error;
+  },
+};
+
+jest.mock("react-intl", () => {
+  const reactIntl = jest.requireActual("react-intl");
+  const intl = reactIntl.createIntl(MOCK_INTL_CONFIG);
+
+  return {
+    ...reactIntl,
+    useIntl: () => intl,
   };
 });